Move init code into app/config/index.js from app/index.js
