+++ client/app/config/styles.js
@@ -1,0 +1,11 @@
+import 'material-design-iconic-font/dist/css/material-design-iconic-font.css';
+import 'font-awesome/css/font-awesome.css';
+import 'ui-select/dist/select.css';
+import 'angular-toastr/dist/angular-toastr.css';
+import 'angular-resizable/src/angular-resizable.css';
+import 'angular-gridster/dist/angular-gridster.css';
+import 'pace-progress/themes/blue/pace-theme-minimal.css';
+
+import '@/assets/css/superflat_redash.css';
+import '@/assets/css/redash.css';
+import '@/assets/css/main.scss';
